+++ backend/src/main/java/com/morjo/controller/LoginController.java
@@ -59,25 +59,4 @@ public class LoginController {
         session.removeAttribute("userId");
         return ResponseEntity.status(HttpStatus.OK).body("로그인 완료");
     }
-
-    @PostMapping("/join")
-    public ResponseEntity<?> join(@RequestBody User user, HttpSession session) {
-        Long kakaoId = (Long) session.getAttribute("kakaoId");
-
-        if (kakaoId == null) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("가입중인 유저가 아닙니다");
-        }
-
-        user.setKakaoId(kakaoId);
-        boolean success = userSerivce.join(user);
-
-        // !TODO 따로 에러를 줘야하는 상황 : 닉네임 중복
-        if (success) {
-            session.removeAttribute("kakaoId");
-            session.setAttribute("userId", user.getUserId());
-            ResponseEntity.status(HttpStatus.CREATED).build();
-        }
-
-        return ResponseEntity.status(HttpStatus.CONFLICT).body("중복된 닉네임입니다");
-    }
 }
