+++ backend/src/main/java/com/morjo/controller/UserController.java
@@ -1,0 +1,39 @@
+package com.morjo.controller;
+
+import com.morjo.model.dto.User;
+import com.morjo.model.service.UserSerivce;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+public class UserController {
+
+    private final UserSerivce userSerivce;
+
+    @PostMapping("/join")
+    public ResponseEntity<?> join(@RequestBody User user, HttpSession session) {
+        Long kakaoId = (Long) session.getAttribute("kakaoId");
+
+        if (kakaoId == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("가입중인 유저가 아닙니다");
+        }
+
+        user.setKakaoId(kakaoId);
+        boolean success = userSerivce.join(user);
+
+        // !TODO 따로 에러를 줘야하는 상황 : 닉네임 중복
+        if (success) {
+            session.removeAttribute("kakaoId");
+            session.setAttribute("userId", user.getUserId());
+            ResponseEntity.status(HttpStatus.CREATED).build();
+        }
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("중복된 닉네임입니다");
+    }
+}
